+++ MCMaps/Models/Document/CartographyMap.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 /// A representation of the basic Minecraft world map.
-struct CartographyMap: Codable, Hashable {
+struct CartographyMap: Codable, Hashable, Sendable {
     /// The seed used to generate the world in-game.
     var seed: Int64