+++ MCMaps/Views/DocumentLaunchView/DocumentLaunchView+Desktop.swift
@@ -1,68 +1,13 @@
 //
-//  DocumentLaunchView.swift
+//  DocumentLaunchView+Desktop.swift
 //  MCMaps
 //
-//  Created by Marquis Kurt on 01-03-2025.
+//  Created by Marquis Kurt on 15-03-2025.
 //
 
 import SwiftUI
 
-#if os(iOS)
-    /// A launch view that lets players create and open files easily.
-    ///
-    /// This will generally be displayed when the app is first loaded.
-    @available(iOS 18.0, *)
-    struct DocumentLaunchView: Scene {
-        @State private var creationContinuation: CheckedContinuation<CartographyMapFile?, any Error>?
-        /// Whether the creation window should be visible.
-        ///
-        /// Applicable to iOS and iPadOS.
-        @Binding var displayCreationWindow: Bool
-
-        /// The proxy map used to create a file temporarily.
-        ///
-        /// Applicable to iOS and iPadOS.
-        @Binding var proxyMap: CartographyMap
-
-        var body: some Scene {
-            DocumentGroupLaunchScene {
-                NewDocumentButton("Create Map", for: CartographyMapFile.self) {
-                    try await withCheckedThrowingContinuation { continuation in
-                        self.creationContinuation = continuation
-                        self.displayCreationWindow = true
-                    }
-                }
-                .sheet(isPresented: $displayCreationWindow) {
-                    NavigationStack {
-                        MapCreatorForm(worldName: $proxyMap.name, mcVersion: $proxyMap.mcVersion, seed: $proxyMap.seed)
-                            .navigationTitle("Create Map")
-                            .toolbar {
-                                ToolbarItem(placement: .confirmationAction) {
-                                    Button("Create") {
-                                        creationContinuation?.resume(returning: .init(map: proxyMap))
-                                        creationContinuation = nil
-                                        displayCreationWindow = false
-                                    }
-                                }
-
-                                ToolbarItem(placement: .cancellationAction) {
-                                    Button("Cancel") {
-                                        creationContinuation?.resume(throwing: CocoaError(CocoaError.userCancelled))
-                                        creationContinuation = nil
-                                        displayCreationWindow = false
-                                    }
-                                }
-                            }
-                    }
-                }
-            } background: {
-                Image(.packMcmeta)
-                    .resizable()
-                    .aspectRatio(contentMode: .fill)
-            }
-        }
-    }
-#else
+#if os(macOS)
     /// A launch view that lets players create and open files easily.
     ///
     /// This will generally be displayed when the app is first loaded.
@@ -82,6 +27,7 @@ import SwiftUI
         @Environment(\.openDocument) private var openDocument
 
         @State private var selectedFile: URL?
+        @State private var displayCreationSheet = false
 
         @ScaledMetric private var fileHeight = 36.0
         @ScaledMetric private var filePaddingH = 4.0
@@ -123,6 +69,25 @@ import SwiftUI
                 .frame(width: 600, height: 400)
                 .toolbarVisibility(.hidden, for: .windowToolbar)
                 .containerBackground(.thinMaterial, for: .window)
+                .sheet(isPresented: $displayCreationSheet) {
+                    NavigationStack {
+                        MapCreatorForm(worldName: $proxyMap.name, mcVersion: $proxyMap.mcVersion, seed: $proxyMap.seed)
+                            .navigationTitle("Create Map")
+                            .formStyle(.grouped)
+                            .toolbar {
+                                ToolbarItem(placement: .confirmationAction) {
+                                    Button("Create") {
+                                        createDocument()
+                                    }
+                                }
+                                ToolbarItem(placement: .cancellationAction) {
+                                    Button("Cancel", role: .cancel) {
+                                        displayCreationSheet = false
+                                    }
+                                }
+                            }
+                    }
+                }
             }
             .windowStyle(.hiddenTitleBar)
             .windowResizability(.contentSize)
@@ -157,8 +122,7 @@ import SwiftUI
                 Spacer()
                 VStack(alignment: .leading) {
                     Button {
-                        newDocument(contentType: .cartography)
-                        dismissWindow(id: "launch")
+                        displayCreationSheet.toggle()
                     } label: {
                         HStack {
                             Label("Create a Map", systemImage: "document.badge.plus")
@@ -251,6 +215,19 @@ import SwiftUI
             }
         }
 
+        private func createDocument() {
+            proxyMap.pins.append(CartographyMapPin(position: .zero, name: "Spawn Point"))
+            proxyMap.recentLocations?.append(.zero)
+            let newFile = CartographyMapFile(map: proxyMap)
+            newDocument(newFile)
+            displayCreationSheet = false
+            Task {
+                // NOTE(alicerunsonfedora): WTF is this bullshit? Can't you just dismiss normally you dirty fuck?
+                try await Task.sleep(nanoseconds: 1000)
+                dismissWindow(id: "launch")
+            }
+        }
+
         private func openDocument(at url: URL) {
             Task {
                 do {
@@ -289,40 +266,18 @@ import SwiftUI
             NSWorkspace.shared.activateFileViewerSelecting([url])
         }
     }
-#endif
 
-private struct DocumentLaunchViewButtonModifier: ViewModifier {
-    func body(content: Content) -> some View {
-        content
-            .frame(maxWidth: .infinity)
-            .padding(.horizontal, 16)
-            .padding(.vertical, 10)
-            .background(Color.secondary.opacity(0.1))
-            .controlSize(.extraLarge)
-            .buttonStyle(.borderless)
-            .bold()
-            .clipped()
-            .clipShape(.rect(cornerRadius: 10))
-    }
-}
+    #if DEBUG
+        extension DocumentLaunchView {
+            var testHooks: TestHooks { TestHooks(target: self) }
 
-extension View {
-    fileprivate func documentLaunchViewButtonStyle() -> some View {
-        self.modifier(DocumentLaunchViewButtonModifier())
-    }
-}
+            struct TestHooks {
+                private let target: DocumentLaunchView
 
-#if DEBUG
-    extension DocumentLaunchView {
-        var testHooks: TestHooks { TestHooks(target: self) }
+                fileprivate init(target: DocumentLaunchView) {
+                    self.target = target
+                }
 
-        struct TestHooks {
-            private let target: DocumentLaunchView
-
-            fileprivate init(target: DocumentLaunchView) {
-                self.target = target
-            }
-            #if os(macOS)
                 @available(macOS 15.0, *)
                 func friendlyUrl(_ url: URL, for currentUser: String) -> String {
                     target.friendlyUrl(url, for: currentUser)
@@ -337,12 +292,7 @@ extension View {
                 func sanitize(url: URL) -> String {
                     target.sanitize(url: url)
                 }
-            #else
-                @available(iOS 18.0, *)
-                var creationContinuation: CheckedContinuation<CartographyMapFile?, any Error>? {
-                    target.creationContinuation
-                }
-            #endif
+            }
         }
-    }
+    #endif
 #endif